split openshift-launch.sh to configure/launch
We need to call the old openshift-launch script to perform its
work and transform the XML file so it is well-formed. We can then
look at parsing it and transforming in the new runtime Python CCT
module (but it must be well-formed first, so the substitutions the
shellscripts perform need to have taken place).

The "exec" line goes to openshift-launch.sh, which is the image's
CMD, the rest to openshift-configure.sh. The exec line needs to
reference variables set in -configure.sh. As a temporary measure,
write those out to /tmp/launch_envs, and source that from -launch.sh.